requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
openai==1.3.5
mcp==1.0.0
python-dotenv==1.0.0
pydantic==2.41.5
pydantic-core==2.41.5
pydantic-settings==2.1.0
asyncpg==0.29.0
better-auth==0.0.1b11
psycopg2-binary==2.9.11
PyJWT==2.11.0
bcrypt==5.0.0
